Simplified implementation
