+++ apps/api/src/app.module.ts
@@ -3,7 +3,6 @@ import { APP_GUARD } from '@nestjs/core';
 import { ThrottlerModule, ThrottlerGuard } from '@nestjs/throttler';
 import { ConfigModule } from './config/config.module';
 import { ConnectionsModule } from './connections/connections.module';
-import { VersionCheckModule } from './version-check/version-check.module';
 import { HealthModule } from './health/health.module';
 import { MetricsModule } from './metrics/metrics.module';
 import { AuditModule } from './audit/audit.module';
@@ -67,7 +66,6 @@ const baseImports = [
     limit: 10000, // Very high default - endpoint-specific limits provide actual rate limiting
   }]),
   ConnectionsModule, // Must come early - provides ConnectionRegistry globally
-  VersionCheckModule, // Global module for version update checks
   HealthModule,
   MetricsModule,
   AuditModule,
