--- apps/api/src/version-check/index.ts
@@ -1,2 +1,0 @@
-export { VersionCheckModule } from './version-check.module';
-export { VersionCheckService } from './version-check.service';
